more bug fixes and refactoring

+++ freighter/path.py
@@ -1,6 +1,6 @@
 from functools import cache, cached_property
 from glob import glob
-from os import getcwd, makedirs
+from os import getcwd, makedirs, remove
 from os.path import expanduser, expandvars, isdir, isfile, realpath, abspath
 from pathlib import PurePosixPath, PureWindowsPath
 from platform import system
@@ -12,8 +12,6 @@ from freighter.exceptions import FreighterException
 
 
 class Path(PureWindowsPath):
-
-
     @classmethod
     @property
     def home(cls):
@@ -98,19 +96,17 @@ class Path(PureWindowsPath):
         if not self.parts:
             return ""
         if self.drive:
-            return "/".join(self.parts).replace("//","/")
+            return "/".join(self.parts).replace("//", "/")
         return "/".join(self.parts)
 
 
-
 class DirectoryPath(Path):
-
     def exists(self) -> bool:
         if isdir(self):
             Console.print(f'{ORANGE}Directory Found "{self}"!', PrintType.VERBOSE)
             return True
         else:
-            Console.print(f'The folder "{self}" does not exist', PrintType.WARN)  # relative to the cwd "{getcwd()}"')
+            Console.print(f'The folder "{self}" does not exist', PrintType.INFO)  # relative to the cwd "{getcwd()}"')
             return False
 
     def assert_exists(self) -> bool:
@@ -119,17 +115,22 @@ class DirectoryPath(Path):
         else:
             raise FreighterException(f'The folder "{self}" does not exist')  # relative to the cwd "{getcwd()}"')
 
-    def rmtree(self):
-        rmtree(self)
+    def delete(self, ask_confirm: bool = False):
+        if not self.exists():
+            return
+        if ask_confirm and input(f'Confirm deletion of directory "{self}"?\nType "yes" to confirm:\n') == "yes":
+            rmtree(self)
+        
 
     def find_files(self, *extensions: str, recursive=False):
         globbed = list[str]()
         if extensions:
             for extension in extensions:
-                globbed += glob(f"{self}/**/**{extension}", recursive=recursive)
+                globstr = f"{self}/**/*{extension}"
+                globbed += glob(globstr, recursive=recursive,)
         else:
-            globbed = glob(f"{self}/**/**", recursive=recursive)
-
+            globbed = glob(f"{self}/*", recursive=recursive)
+        
         result = list[FilePath]()
         for path in globbed:
             result.append(FilePath(path))
@@ -157,7 +158,6 @@ class DirectoryPath(Path):
 
 
 class FilePath(Path):
-
     def exists(self) -> bool:
         if isfile(self):
             Console.print(f'{ORANGE}File Found "{self}"!', PrintType.VERBOSE)
@@ -165,6 +165,13 @@ class FilePath(Path):
         else:
             Console.print(f'The file "{self}" does not exist', PrintType.WARN)  # relative to the cwd "{getcwd()}"')
             return False
+
+    def delete(self, ask_confirm: bool = False):
+        if not self.exists():
+            return
+        if ask_confirm and input(f'Confirm deletion of file "{self}"?\nType "yes" to confirm:\n') == "yes":
+            remove(self)
+        
 
     def assert_exists(self):
         if self.exists():
